Assign each attribute to separate list if multiple values
If a file contains multiple activities (like views), we use a dict to
capture them. Before this commit, we directly assigned the dict to an
attribute of the FacebookQuantifier class (e.g. self.messages). This
made creating the spreadsheet and summarizing the data more complicated.
To simplify the code, we now check the returned dict and assign each
key-value pair to a separate attribute of the class. So instead of
`self.messages['sent_message']` we just have `self.sent_message`.