Remove Pandas dependency, add verbose flag
Pandas was used only to create the CSV spreadsheet, which seems like
an unnecessary huge dependency for this one purpose. Now we use the csv
module from Python's standard library.

As a minor additional change, the script now only prints an overview
of the data gathered when the new `--verbose`/`-v` flag is used.